+++ Tests/Functional/TestBundle/Entity/Product.php
@@ -1,9 +1,24 @@
 <?php
 
+/*
+ * This file is part of the JsFormValidationBundle.
+ *
+ * (c) Abhoryo <abhoryo@free.fr>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace APY\JsFormValidationBundle\Tests\Functional\TestBundle\Entity;
 
 use Symfony\Component\Validator\Constraints as Assert;
 
+/**
+ * Product Entity which is used for testing purposes
+ *
+ * @author Vitaliy Demidov   <zend@i.ua>
+ * @since  15 Aug 2012
+ */
 class Product
 {
     /**
@@ -34,6 +49,14 @@ class Product
      * @Assert\NotBlank(message = "Password should not be blank.")
      */
     protected $password;
+
+    /**
+     * This field is used for test property_path FALSE.
+     * Validation constraint for it should not appear in the script.
+     *
+     * @Assert\NotBlank()
+     */
+    protected $excluded;
 
 
     /**
@@ -93,4 +116,15 @@ class Product
     {
         $this->password = $password;
     }
+
+    public function getExcluded()
+    {
+        return $this->excluded;
+    }
+
+    public function setExcluded($excluded)
+    {
+        $this->excluded = $excluded;
+    }
+
 }